Preparing to turn more code into utility methods
Probably going to code in a style where all logic code for determining block placement will have their own helper/util classes instead of being in Item classes. This makes items more replaceable. Also added classes for IBitType, going to finish that later.

+++ src/main/java/nl/dgoossens/chiselsandbits2/api/modes/ChiselModeManager.java
@@ -7,25 +7,19 @@ import nl.dgoossens.chiselsandbits2.network.NetworkRouter;
 import nl.dgoossens.chiselsandbits2.network.packets.PacketSetChiselMode;
 
 public class ChiselModeManager {
-    public static void changeChiselMode(
-            final ItemMode newMode)
-    {
+    public static void changeChiselMode(final ItemMode newMode) {
         final PacketSetChiselMode packet = new PacketSetChiselMode(newMode);
         NetworkRouter.sendToServer( packet );
     }
 
-    public static void scrollOption(
-            final ItemMode.Type tool,
-            ItemMode currentMode,
-            final int dwheel )
-    {
+    public static void scrollOption(final ItemMode.Type tool, ItemMode currentMode, final double dwheel) {
         int offset = currentMode.ordinal();
         do {
             offset = testOffset(offset + ( dwheel < 0 ? -1 : 1 ));
             currentMode = ItemMode.values()[offset];
         } while(currentMode.getType()!=tool);
 
-        changeChiselMode( currentMode );
+        changeChiselMode(currentMode);
     }
 
     private static int testOffset(int offset) {
